pyproject.toml
[project]
name = "facebook-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "enkhbold470", email = "enkhbold470@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "flask==3.1.0",
    "flask-login==0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn==23.0.0",
    "pg8000>=1.30.5",
    "psycopg2-binary>=2.9.10",
    "python-dotenv==1.0.1",
    "sqlalchemy>=2.0.41",
    "waitress==3.0.2",
    "werkzeug==3.1.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
